Placed test_models into test dir
